+++ changelog.md
@@ -1,3 +1,5 @@
 # Changelog of Knockout Crud
 
 1.0.0 Initial commit
+1.1.0 New observable property itemForAdding
+1.2.0 New method removeItem to remove an instance when editing that object